prepared ResourceListView and StorageListView

+++ src/main/java/edu/wsiiz/repairshop/storage/domain/storage/Delivery.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Delivery {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int deliveryId;
+    private Long deliveryId;
 
     @Temporal(TemporalType.DATE)
     private Date deliveryDate;
